完善文档说明

+++ app/src/main/java/com/kingsun/thirddock/MainActivity.kt
@@ -215,7 +215,7 @@ class MainActivity : AppCompatActivity() , EasyPermissions.PermissionCallbacks, 
         bookResource?.apply {
             val params = JsonObject()
             params.addProperty("Device",Device)
-            params.addProperty("BookId",binding.etBookId.text.toString())
+            params.addProperty("BookId",binding.etBookId.text.toString().toInt())
             savePath?.let {
                 params.addProperty("ResourcePath",it)
             }
